Add module CommandUIOptParse
